Use 1 worker
This is the correct solution. Gunicorn is configured to use an event-
based handler and thus doesn't need multiprocessing. Moreover the
global APScheduler breaks with more schedulers. --preload cannot fix
this because all requests handlers need to put unlock timeouts onto the
scheduler; with --preload all but one scheduler would not run.